+++ NEWS.md
@@ -37,6 +37,7 @@
 * New `load-balancer` option `timeout` limits how long to wait for a response from a node, defaults to 10 seconds. (Pro #1642)
 * New `load-balancer` option `verify-ssl-certs` for testing nodes with self-signed certificates when using SSL. (Pro #1504)
 * New `launcher-verify-ssl-certs` and `launcher-sessions-callback-verify-ssl-certs` options for testing with self-signed certificates when using SSL. (Pro #1504)
+* R sessions can now be renamed from within the session or the home page. (Pro #1572)
 
 ### Bugfixes